+++ pkg/server/exec/handlers.go
@@ -1,14 +1,19 @@
 package exec
 
 import (
+	"time"
+
 	"github.com/luizalabs/teresa/pkg/goutil"
 	execpb "github.com/luizalabs/teresa/pkg/protobuf/exec"
 	"github.com/luizalabs/teresa/pkg/server/database"
 	"google.golang.org/grpc"
 )
 
+const keepAliveMessage = "\u200B" // Zero width space
+
 type Service struct {
-	ops Operations
+	ops              Operations
+	keepAliveTimeout time.Duration
 }
 
 func (s *Service) Command(req *execpb.CommandRequest, stream execpb.Exec_CommandServer) error {
@@ -25,6 +30,8 @@ func (s *Service) Command(req *execpb.CommandRequest, stream execpb.Exec_Command
 	var msg string
 	for {
 		select {
+		case <-time.After(s.keepAliveTimeout):
+			msg = keepAliveMessage
 		case err := <-errChan:
 			return err
 		case m, ok := <-cmdMsgs:
@@ -44,6 +51,6 @@ func (s *Service) RegisterService(grpcServer *grpc.Server) {
 	execpb.RegisterExecServer(grpcServer, s)
 }
 
-func NewService(ops Operations) *Service {
-	return &Service{ops: ops}
+func NewService(ops Operations, keepAliveTimeout time.Duration) *Service {
+	return &Service{ops: ops, keepAliveTimeout: keepAliveTimeout}
 }