+++ pkg/server/server.go
@@ -131,12 +131,11 @@ func registerServices(s *grpc.Server, opt Options, uOps user.Operations) error {
 	e.RegisterService(s)
 
 	buildOpts := &build.Options{
-		SlugBuilderImage:   opt.DeployOpt.SlugBuilderImage,
-		SlugRunnerImage:    opt.DeployOpt.SlugRunnerImage,
-		SlugStoreImage:     opt.DeployOpt.SlugStoreImage,
-		BuildLimitCPU:      opt.DeployOpt.BuildLimitCPU,
-		BuildLimitMemory:   opt.DeployOpt.BuildLimitMemory,
-		DefaultServiceType: opt.DeployOpt.DefaultServiceType,
+		SlugBuilderImage: opt.DeployOpt.SlugBuilderImage,
+		SlugRunnerImage:  opt.DeployOpt.SlugRunnerImage,
+		SlugStoreImage:   opt.DeployOpt.SlugStoreImage,
+		BuildLimitCPU:    opt.DeployOpt.BuildLimitCPU,
+		BuildLimitMemory: opt.DeployOpt.BuildLimitMemory,
 	}
 	bOps := build.NewBuildOperations(opt.Storage, appOps, execOps, opt.K8s, buildOpts)
 	b := build.NewService(bOps, opt.DeployOpt.KeepAliveTimeout)