+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [NEXT_RELEASE]
+### Changed
+- [client] Print app name and current cluster on command `app delete`
+
 ## [0.22.0] - 2018-06-07
 ### Changed
 - [server] Update client-go to v6.0.0 and bump minimum k8s version to 1.9