+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [NEXT_RELEASE]
 ### Added
-- [client] Add support to remoVe remove cluster from configuration file
+- [client] Add support to remove remote cluster from configuration file
 
 ### Changed
 - [client] Print app name and current cluster on command `app delete`