+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - [client] Print app name and current cluster on command `app delete`
+- [server] Set default backoff limit to 3 on cronjobs
 
 ### Fixed
 - [server] Fix cloud provider detection on gce