+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - [client] Print app name and current cluster on command `app delete`
 - [server] Set default backoff limit to 3 on cronjobs
 - [server] Always use service type LoadBalancer on builds
+- [server] Use app.tgz as source code tarball name
 
 ### Fixed
 - [server] Fix cloud provider detection on gce