+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [NEXT_RELEASE]
+## [0.23.0] - 2018-06-21
 ### Added
-- [client] Add support to remove remote cluster from configuration file
+- [client] Support to remove remote cluster from configuration file
 
 ### Changed
 - [client] Print app name and current cluster on command `app delete`